refactor(tui): extract shade map + export JSON as pure functions
app.rs was ~2247 lines, half tests half 32-field TUI state machine.
An architect review rejected a full App-struct split (handle_key_event
touches 17 fields — the coupling is essential to a TUI reducer) but
approved moving two genuinely-pure computations out:

- `build_model_shade_map(&[ModelUsage]) -> HashMap<String, Color>`
  moves to `tui/colors.rs`. Doesn't need any App state beyond `data.models`.
- `build_export_json(&UsageData) -> Result<String>` moves to
  `tui/export.rs`. File I/O and the status-message side effect stay
  on `App` where they belong.

`App::build_model_shade_map` and `App::export_to_json` keep the same
signatures as thin call-site wrappers so nothing external shifts. The
existing shade-map unit tests still go through `App::build_model_shade_map`
and are unchanged — though they could now target the pure function
directly if a future pass wants to drop the `App` setup.

Net: app.rs 2247 -> 2174 (-73). Not huge, but the two functions that
moved are the ones that actually wanted to be pure — the state-machine
methods that remain are where they belong.

Constraint: don't alter CLI/TUI behavior — only relocate pure logic
Rejected: full split into event.rs / sort.rs / dialog.rs | handle_key_event touches 17 fields; splitting just moves the tangle through more indirection
Rejected: move tests to sibling file via `#[path]` | cosmetic, doesn't improve anything
Confidence: high
Scope-risk: narrow
Directive: when adding new pure computations on UsageData, consider tui/colors.rs or tui/export.rs before expanding App's impl block
Not-tested: JSON output byte-equivalence (serde_json::to_string_pretty is used identically); relies on downstream tooling that reads the export to flag regressions
