+++ crates/tokscale-cli/src/tui/app.rs
@@ -14,7 +14,7 @@ use super::data::{AgentUsage, DailyUsage, DataLoader, HourlyUsage, ModelUsage, U
 use super::settings::Settings;
 use super::themes::{Theme, ThemeName};
 use super::ui::dialog::{ClientPickerDialog, DialogStack};
-use super::ui::widgets::{get_model_color, get_provider_from_model, get_provider_shade};
+use super::ui::widgets::get_model_color;
 
 /// Configuration for TUI initialization
 pub struct TuiConfig {
@@ -286,34 +286,7 @@ impl App {
     }
 
     pub fn build_model_shade_map(&mut self) {
-        self.model_shade_map.clear();
-
-        // Aggregate cost per (provider, model_name) so that the same model
-        // appearing in multiple group-by buckets (e.g. WorkspaceModel) gets
-        // counted once — preventing rank inflation and last-write-win shade
-        // collisions.
-        let mut provider_models: HashMap<&str, HashMap<&str, f64>> = HashMap::new();
-        for m in &self.data.models {
-            let provider = get_provider_from_model(&m.model);
-            let cost = if m.cost.is_finite() { m.cost } else { 0.0 };
-            *provider_models
-                .entry(provider)
-                .or_default()
-                .entry(m.model.as_str())
-                .or_insert(0.0) += cost;
-        }
-
-        for (provider, models_map) in provider_models {
-            let mut models: Vec<(&str, f64)> = models_map.into_iter().collect();
-            // Primary sort by cost descending, with a stable secondary sort on
-            // the model name so cost ties (including all-zero costs on first
-            // load) produce identical shade assignments across refreshes.
-            models.sort_by(|a, b| b.1.total_cmp(&a.1).then_with(|| a.0.cmp(b.0)));
-            for (rank, (model_name, _)) in models.iter().enumerate() {
-                let color = get_provider_shade(provider, rank);
-                self.model_shade_map.insert(model_name.to_string(), color);
-            }
-        }
+        self.model_shade_map = super::colors::build_model_shade_map(&self.data.models);
     }
 
     pub fn model_color(&self, model: &str) -> Color {
@@ -869,59 +842,12 @@ impl App {
     }
 
     fn export_to_json(&mut self) {
-        let export_data = serde_json::json!({
-            "models": self.data.models.iter().map(|m| serde_json::json!({
-                "model": m.model,
-                "provider": m.provider,
-                "client": m.client,
-                "tokens": {
-                    "input": m.tokens.input,
-                    "output": m.tokens.output,
-                    "cacheRead": m.tokens.cache_read,
-                    "cacheWrite": m.tokens.cache_write,
-                    "total": m.tokens.total()
-                },
-                "cost": m.cost,
-                "sessionCount": m.session_count
-            })).collect::<Vec<_>>(),
-            "agents": self.data.agents.iter().map(|a| serde_json::json!({
-                "agent": a.agent,
-                "clients": a.clients,
-                "tokens": {
-                    "input": a.tokens.input,
-                    "output": a.tokens.output,
-                    "cacheRead": a.tokens.cache_read,
-                    "cacheWrite": a.tokens.cache_write,
-                    "total": a.tokens.total()
-                },
-                "cost": a.cost,
-                "messageCount": a.message_count
-            })).collect::<Vec<_>>(),
-            "daily": self.data.daily.iter().map(|d| serde_json::json!({
-                "date": d.date.to_string(),
-                "tokens": {
-                    "input": d.tokens.input,
-                    "output": d.tokens.output,
-                    "cacheRead": d.tokens.cache_read,
-                    "cacheWrite": d.tokens.cache_write,
-                    "total": d.tokens.total()
-                },
-                "messageCount": d.message_count,
-                "turnCount": d.turn_count,
-                "cost": d.cost
-            })).collect::<Vec<_>>(),
-            "totals": {
-                "tokens": self.data.total_tokens,
-                "cost": self.data.total_cost
-            }
-        });
-
         let filename = format!(
             "tokscale-export-{}.json",
             chrono::Utc::now().format("%Y%m%d-%H%M%S")
         );
 
-        match serde_json::to_string_pretty(&export_data) {
+        match super::export::build_export_json(&self.data) {
             Ok(json) => match std::fs::write(&filename, json) {
                 Ok(_) => self.set_status(&format!("Exported to {}", filename)),
                 Err(e) => self.set_status(&format!("Export failed: {}", e)),
@@ -1093,6 +1019,7 @@ impl App {
 
 #[cfg(test)]
 mod tests {
+    use super::super::ui::widgets::get_provider_shade;
     use super::*;
     use crate::tui::data::{ModelUsage, TokenBreakdown};
