+++ crates/tokscale-cli/src/tui/mod.rs
@@ -1,9 +1,11 @@
 mod app;
 mod cache;
 pub mod client_ui;
+mod colors;
 pub mod config;
 pub mod data;
 mod event;
+mod export;
 pub mod settings;
 mod themes;
 mod ui;
